Bug Fix: The sound is missing on some notifications
This was due to our previous changes to improve notifications on devices running API level < 24.
Indeed the summary was preventing the user from seeing the messages when only one room is notified on devices running API level < 24.
We apply here another patch to display the unique summary line, instead of "1 Notification"